requirements.txt
# Updated requirements.txt with compatible versions
flask==3.0.2
flask-cors==4.0.0
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4  # Compatible with TensorFlow 2.18.0
Pillow==10.3.0
tensorflow-intel==2.18.0